`@kbn/react-env` for easier access of essential kibana env properties (#243813)
## Summary

On the client-side we sometime need to make small changes, like updating
a tooltip or changing a query parameter based on the environment (ech,
local, serverless, kibanaVersion). For example, we might need to use
different text for a tour or adjust the query parameter for the feedback
URL.

Depending on how deep in the UI code the change needs to be done, we
jump through various the hoops doing this. adding more props in
component tree and services etc, etc. **What seems to be an hour of work
becomes a messy and annoying 2-3 day task, or we decide it's not worth
it in the end.**

This package attempts to solve this by providing an easier access to
essential properties from `CoreContext['env']` object in React
components.

This should be used with caution for small tweaks. Shouldn't be used for
swapping the whole pages.

### DX

#### Access

```ts
import { useIsServerless } from '@kbn/react-env';

export const MyComponent = () => {
  const isServerless = useIsServerless();
  return <div>{isServerless ? 'Serverless' : 'Stateful'}</div>;
};
```

#### Requirements

The hooks can be used anywhere where Kibana render context provider is
configured correctly.

```
ReactDOM.render(
    core.rendering.addContext(<MyApp/>)
) 
```


> [!CAUTION]
> That env checks should still be used with caution! It's just that if
we would otherwise propagate props from core context, we can now use it
more cleanly.

---------

Co-authored-by: kibanamachine <42973632+kibanamachine@users.noreply.github.com>