chore(slo): Fix flicking list and welcome page when no SLO definition exists (#243923)
Fix https://github.com/elastic/kibana/issues/239679

## Summary

Currently when the user has no SLO defined on the active space, clicking
on the SLO page would attempt at rendering the SLO list page, and
immediately redirecting to the SLO welcome page with a UI flicker.

This PR relies on the SLO Definitions API to determine if the user
should be redirected:
- to the SLO welcome page from the SLO list page
- to the SLO list page from the SLO welcome page. 
- to the SLO welcome page from the SLO management page

The SLO Definition API has been updated to allow `perPage: 0` as this is
useful to get a count without returning any SLO definitions. Previously
overwritten with a fallback value of 100.

**How to test**
1. Go to the SLO page, assert the welcome page is shown
2. Go to the SLO Management page, assert the welcome page is shown
3. Create an SLO, assert the SLO List page is shown

Known issue that is going to be fixed with another PR: the health
callout is shown right after creating an SLO due to the transform
starting.