[Agent Builder] Fix "forced" handover to answering step (#244091)
## Summary

Follow-up of https://github.com/elastic/kibana/pull/241607
Fix https://github.com/elastic/search-team/issues/11720 (hopefully)

There was a bug in https://github.com/elastic/kibana/pull/241607 which
was causing forced handover to not be visible in the answering agent's
message history, causing the answering agent to still try to call tools.

This PR fixes the bug, which - from my testing - seems to be sufficient
to fix the behavior in "forced handover" mode.

### Traces

Before the fix:
https://35-187-109-62.sslip.io/projects/UHJvamVjdDoyNTM=/spans/8726d789d5d20d5019f8e296c39ee21f?selectedSpanNodeId=U3BhbjoxMzI3MDQ1
After the fix:
https://35-187-109-62.sslip.io/projects/UHJvamVjdDoyNTM=/spans/cb24629bd37c7b93091668223891f1c5?selectedSpanNodeId=U3BhbjoxMzI3MDY5

---------

Co-authored-by: kibanamachine <42973632+kibanamachine@users.noreply.github.com>
Co-authored-by: Elastic Machine <elasticmachine@users.noreply.github.com>