[242737][Synthetics] Fetch trend data only from visible monitors in the overview page (#243682)
## Summary
 closes #242737 

Update the `useOverviewTrendsRequests` hook to fetch trend data only for
those monitors that are or has been visible in the screen.

Before, the `useOverviewTrendsRequests` hook would fetch data for all
the existing monitors. Now, it gets an incremental slice from the
monitors depending on the number of rows and the next page that needs to
be fetched. This logic is very much the same that was being used some
time ago but that it was removed in this
[PR](https://github.com/elastic/kibana/pull/219060)

* Before: trend data is fetched for all the existing monitors at once

![demoFixTrend1](https://github.com/user-attachments/assets/092ae7a9-d1eb-4b20-bc96-ad12344d120b)


* After: tr

![demoFixTrend2](https://github.com/user-attachments/assets/d87fa860-5463-441c-853f-cce961aafb31)
end data is fetched as the use scrolls down






### Checklist

Check the PR satisfies following conditions. 

Reviewers should verify this PR satisfies this list as well.


- [x] [Unit or functional
tests](https://www.elastic.co/guide/en/kibana/master/development-tests.html)
were updated or added to match the most common scenarios



- [ ] The PR description includes the appropriate Release Notes section,
and the correct `release_note:*` label is applied per the
[guidelines](https://www.elastic.co/guide/en/kibana/master/contributing.html#kibana-release-notes-process)
- [ ] Review the [backport
guidelines](https://docs.google.com/document/d/1VyN5k91e5OVumlc0Gb9RPa3h1ewuPE705nRtioPiTvY/edit?usp=sharing)
and apply applicable `backport:*` labels.

---------

Co-authored-by: Francesco Fagnani <francesco.fagnani@elastic.co>
Co-authored-by: Francesco Fagnani <fagnani.francesco@gmail.com>