[Streams] Include failure store disabled option (#243658)
## Summary
In https://github.com/elastic/kibana/pull/242295/ we introduced the
logic for handling disabled failure store in the backend and display it
in the cards but the changes for enabling it in the modal were missing.
This PR introduces that.

Note: Failure Store disabled option is not available in serverless

<img width="618" height="410" alt="Screenshot 2025-11-25 at 16 22 21"
src="https://github.com/user-attachments/assets/1bea6b8b-2b2b-41d9-9111-a0ccba815b19"
/>

The modal now includes an optional prop to name the "disabled" button,
since Index Management uses the "disabled" concept and Streams
"Indefinite". Also, reorder the buttons in the modal for consistency
with Streams and not important for Index Management.