feat(security, uiam): make `@kbn\es` to manage UIAM containers when in UIAM mode (#243752)
## Summary

Make `@kbn\es` to manage UIAM containers when in Serverless + UIAM mode
as it's going to be a default for Serverless at some point.

__NOTE:__ This PR includes some groundwork for initializing UIAM in
tests (i.e., passing the `--uiam` flag from test configs down to
`@kbn/es` APIs). However, it doesn't include tests yet. The reason is
that I'd like to start with Scout tests, but it's not yet possible to
have a dedicated Scout config where we can toggle UIAM mode. I'm working
with @elastic/appex-qa to add the remaining changes to make this happen,
but it will likely go in a follow-up. If you feel like we need to move
the test-related changes to a separate PR, just let me know.

## How to test

First, run Elasticsearch Severless in the UIAM mode:
```shell
$ yarn es serverless --projectType oblt --uiam
---
   │ info es01 is running.
   │        Container Name: es01
   │        Container Id:   9bb06af54a4b98c4dd3f75e21e6b1a47e5bad7cf2b8e0d7f18e3ec0e3321cf72
   │
   │        View logs:            docker logs -f es01
   │        Shell access:         docker exec -it es01 /bin/bash
   │
   │ info es02 is running.
   │        Container Name: es02
   │        Container Id:   688e43ce813c68450c70c87f2ccc29738e06e0a45a7ccd4041c86718ac103866
   │
   │        View logs:            docker logs -f es02
   │        Shell access:         docker exec -it es02 /bin/bash
   │
   │ info es03 is running.
   │        Container Name: es03
   │        Container Id:   658fe85658de8d0915e63480dc6b96b1e5e4e092e48314b001965c039eb104f4
   │
   │        View logs:            docker logs -f es03
   │        Shell access:         docker exec -it es03 /bin/bash
...
 info Waiting for "uiam-cosmosdb" container (starting)…
 info Waiting for "uiam" container (starting)…
   │ info The "uiam-cosmosdb" container is up and ready.
   │        Container Name: uiam-cosmosdb
   │        Container Id:   e0ea6a43bfc5379b10cf5af941b6677e0cb8d90e075b4b36ff0c487ff188386a
   │
   │        View logs:            docker logs -f uiam-cosmosdb
   │        Shell access:         docker exec -it uiam-cosmosdb /bin/bash
   │
   │ info The "uiam" container is up and ready.
   │        Container Name: uiam
   │        Container Id:   b4a9b5f56bc2155201fc942e22d1fde3d51453e60b45cf74693b43535edd1cb4
   │
   │        View logs:            docker logs -f uiam
   │        Shell access:         docker exec -it uiam /bin/bash
   │
...
 succ ES cluster is ready
 succ Created role mapping for mock identity provider. You can now login using cloud-saml-kibana realm
```

Then, run Kibana Serverless in the UIAM mode:
```shell
$ yarn start --serverless=oblt --uiam
---
...
Kibana will be configured to support UIAM.
[2025-11-26T09:17:56.417+01:00][INFO ][root] Kibana is starting
...
[2025-11-26T09:19:34.241+01:00][INFO ][status] Kibana is now available
```

At this point, you should be able to log in into Kibana through a mock
IdP provider/select as you'd normally do.

---------

Co-authored-by: kibanamachine <42973632+kibanamachine@users.noreply.github.com>