[Streams] Improve error messaging when expensive queries are disabled in Streams schema editor (#243406)
Closes #243161 

## Summary

This PR enhances the user experience in the Streams schema editor when
the Elasticsearch cluster setting `search.allow_expensive_queries` is
set to false. Instead of showing a raw, technical error message, users
now see a user-friendly warning that explains the situation and allows
them to proceed with caution.


## Problem

When `search.allow_expensive_queries `is disabled on an Elasticsearch
cluster, the Streams schema editor's field simulation fails because it
relies on runtime mappings.

## Testing

On the DevTools, set the `search.allow_expensive_queries` to false

```bash
PUT /_cluster/settings
{
  "transient": { "search.allow_expensive_queries": false }
}
```

You can make sure it is set to false by querying
`GET _cluster/settings`

Or by checking the logs of the Elasticsearch cluster running
`info [o.e.c.s.ClusterSettings] [YourMachine] updating
[search.allow_expensive_queries] from [true] to [false]`

Once it is set to false:

**Schema Editor**

- Navigate to Streams and open a stream (not a wired stream)
- Make schema changes (add/modify fields)
- Add field and submit changes - should see the new user-friendly
warning message
- Verify the "Confirm changes" button is enabled
- Click to proceed - changes should be applied successfully

**Processing**

- Navigate to Streams and open a stream (not a wired stream)
- Create some processor that has an expensive query
- Save changes - should see the new user-friendly warning message
- Verify the "Confirm changes" button is enabled
- Click to proceed - changes should be applied successfully


## Unhandled case

For the global search we also use expensive queries, but there is no
error triggered if the `search.allow_expensive_queries` is set to false.
The difference here is that the wired streams will not show up on the
search results, while they do appear if the
`search.allow_expensive_queries` is set to true.