[Streams 🌊] Warn about outdated documents (#242904)
## Summary

In order to have a better change tracking system for streams, this PR
introduces `updated_at` timestamps for streams documents that are meant
to record the moment in time when changes occur in a stream. There are 2
of this timestamps being introduced. Each one has a different level of
nesting, given that they track changes to specific parts of the whole
stream document.

The first one lives at the root of the stream document (next to name and
description) and tracks changes to any part of the stream, which means
all upsert requests processed for the stream will update it. It is
present in all stream types.

The second one is more specific and only present for ingest streams
(Classic and Wired). It lives in `ingest.processing.updated_at` and
tracks changes exclusively for the processing config of the stream. This
property will be used in the future to display outdated document
warnings in the streams UI while working in the processing tab.

The processing timestamp is then used by the UI to determine wether to
show an outdated documents warning when configuring the processing step.
The warning will be shown when at least 1 document in the current
simulation samples set is older than the most recent stream processing
timestamp.

- The warning will prompt users to refresh the samples in order to get
up to date data. But, it may be possible for the warning not to go away
after refreshing if ingestion is stopped or slow
- The outdated documents logic only applies to latest and KQL samples.
It will never be displayed for custom samples sets given that the
dataset in this case is static thus the concept of being outdated
doesn't really apply.

### Visual Updates 👀

<img width="1365" height="907" alt="Screenshot 2025-11-24 at 12 05 55"
src="https://github.com/user-attachments/assets/405ff623-339c-49ce-aefe-5386cfd9030d"
/>

### Reviewer hints

The most important changes are in the `kbn-streams-schema` package,
mostly around ingest types and model_validation. So please pay a closer
look at the changes there

Other significant changes are in the `streams` plugin, under
`state_management` (where the decision to update the processing
timestamp is made) and `migrate_on_read.ts` to backfill the new
properties in existing streams.

For the UI, the most relevant updates are around the state machines for
data sources and stream processing. The data source machine is
responsible for determining whether the warning should be shown or not,
as those machines are the owners of the data in the samples and also
know the kind of data source that also influences wether the icon should
show.
The stream processing machine also plays its part. Given that it owns
the stream definition, it handles forwarding the latest processing
timestamp to the data source machines via machine-to-machine events

## `terminate_after` Updates
To address the issue detected in
https://github.com/elastic/kibana/pull/242904#pullrequestreview-3515851161
we have decided to remove the `terminate_after` parameter from the
queries we run to fetch documents for the processing simulation. This
will ensure that data sources, other than custom samples, always return
the most up to date samples from the stream.

This could potentially cause performance issues in samples documents
retrieval, given that we no longer have the early termination in shards
that `terminate_after` was providing. Even though queries for samples
might now be slower, we consider that a worthy trade-off in exchange for
getting the most up to date documents and fixing the ever present
warning.

### Telemetry
A new custom EBT event was added,
`streams-processing-simulation-samples-fetch-latency` that will report
statistics on the execution time of simulations samples queries. This
way we can keep an eye out on this and detect wether the query time is
too high.

### Timeout
To prevent the page from hanging indefinitely while waiting for the
query to run, we've added a 10s timeout in the data collector actor for
simulations samples on ES side. If any query take longer than that we
will do a best effort approach to display any samples returned as a
partial result from ES before the query timed out

---------

Co-authored-by: kibanamachine <42973632+kibanamachine@users.noreply.github.com>