[Discover][APM] Update full-screen waterfall flyouts (#244983)
## Summary

Relates to https://github.com/elastic/kibana/issues/244593

With the upcoming release of the new flyout system, we reviewed the
traces experience in the Discover.

Two different issues came up:

**Feature branch w/ the new system**
The custom `z-index` we had in place was no longer applied, so we had to
force the styles using `!important`.
This is only a temporary workaround, and it will be removed once we
migrate the full-screen waterfall to the new system.

|Before|After|
|-|-|

|![eui_branch_before](https://github.com/user-attachments/assets/08ea981b-9d17-4aa7-a26f-f636ea981402)|![eui_branch_after](https://github.com/user-attachments/assets/90dd1cbb-246a-4652-a270-eba65c35256c)|


**Main branch**
The flyout position that was previously set to `top: 0` to keep it
aligned with the top of the page stopped working.
The issue was introduced by the Kibana Layout Changes, which set
`core.chrome.layoutType` to `'grid'`.
When testing with `core.chrome.layoutType` set to `'legacy-fixed'`, the
margin doesn’t appear.

We added a workaround to ensure the flyout is correctly positioned in
full-screen mode.

|Before|After|
|-|-|

|![main_before](https://github.com/user-attachments/assets/ff27d9d8-1bdd-44a5-bffd-1c4d6b6256b8)|![main_after](https://github.com/user-attachments/assets/b32cee7d-9c06-4475-b5a9-3fcf6d02abd3)|


Even though the first point only becomes visible once the new EUI with
the updated system is introduced, adding the fix to main now does no
harm.

---------

Co-authored-by: kibanamachine <42973632+kibanamachine@users.noreply.github.com>