[ConnectorsV2] Allow specifying custom global headers for a single file connector type. (#245132)
## Summary

Updates single file connector auth spec to allow specifying custom
headers that will be added to the configured axios instance for the
connector type.

For our existing connectors, we typically save any header values in the
connector saved object as part of the `config`. For the single file
connectors, I thought it would be more dynamic to add it to the
`ActionType` and read it at execution time. This means that if a
connector spec is updated with a new header value, existing connectors
of that type (stored as saved objects) would use the updated headers the
next time they're executed.