🌊 Fix validation for partially invalid streamlang steps (#245367)
This PR fixes the problem that after adding a condition, it's not
possible to add more processors anymore.

This happened because `isWhereBlockSchema` would return false for a
well-formed where-block if it contains an invalid processor, which is
the case when adding a new processor (because the default grok state has
an empty field name, which is not valid as per the schema). This broke
the validation logic.

By using `isWhereBlock` it makes the check cheaper and also doesn't fail
on invalid children. Also added a test for this.