🌊 Streams: More data quality tests (#245376)
Improves test coverage for the data quality tab.

The data quality tab is kind of special because it's sharing most of its
functionality with the data set quality page, which has a substantial
test suite already. So this PR is making the existing very limited test
suite a bit better but still keep it a smoke test style of test to keep
the redundancy low.

This should give us confidence that the integration is working fine,
with tests for all the details being deferred to the data quality page
tests.